+++ example/src/resolvers/types/typemap.ts
@@ -1,37 +1,38 @@
-import { ITypes } from "../generated/resolvers";
+import { ITypes } from "../../generated/resolvers";
 
-import { QueryParent } from "./Query";
-import { MutationParent } from "./Mutation";
-import { ViewerParent } from "./Viewer";
-import { AuthPayloadParent } from "./AuthPayload";
-import { MutationResultParent } from "./MutationResult";
-import { ExperiencesByCityParent } from "./ExperiencesByCity";
-import { HomeParent } from "./Home";
-import { ReservationParent } from "./Reservation";
-import { ExperienceParent } from "./Experience";
-import { ReviewParent } from "./Review";
-import { NeighbourhoodParent } from "./Neighbourhood";
-import { LocationParent } from "./Location";
-import { PictureParent } from "./Picture";
-import { CityParent } from "./City";
-import { ExperienceCategoryParent } from "./ExperienceCategory";
-import { UserParent } from "./User";
-import { PaymentAccountParent } from "./PaymentAccount";
-import { PlaceParent } from "./Place";
-import { BookingParent } from "./Booking";
-import { NotificationParent } from "./Notification";
-import { PaymentParent } from "./Payment";
-import { PaypalInformationParent } from "./PaypalInformation";
-import { CreditCardInformationParent } from "./CreditCardInformation";
-import { MessageParent } from "./Message";
-import { PricingParent } from "./Pricing";
-import { PlaceViewsParent } from "./PlaceViews";
-import { GuestRequirementsParent } from "./GuestRequirements";
-import { PoliciesParent } from "./Policies";
-import { HouseRulesParent } from "./HouseRules";
-import { AmenitiesParent } from "./Amenities";
+import { QueryParent } from "../Query";
+import { MutationParent } from "../Mutation";
+import { ViewerParent } from "../Viewer";
+import { AuthPayloadParent } from "../AuthPayload";
+import { MutationResultParent } from "../MutationResult";
+import { ExperiencesByCityParent } from "../ExperiencesByCity";
+import { HomeParent } from "../Home";
+import { ReservationParent } from "../Reservation";
+import { ExperienceParent } from "../Experience";
 
-import { Context } from "./types/context";
+import { ReviewParent } from "../Review";
+import { NeighbourhoodParent } from "../Neighbourhood";
+import { LocationParent } from "../Location";
+import { PictureParent } from "../Picture";
+import { CityParent } from "../City";
+import { ExperienceCategoryParent } from "../ExperienceCategory";
+import { UserParent } from "../User";
+import { PaymentAccountParent } from "../PaymentAccount";
+import { PlaceParent } from "../Place";
+import { BookingParent } from "../Booking";
+import { NotificationParent } from "../Notification";
+import { PaymentParent } from "../Payment";
+import { PaypalInformationParent } from "../PaypalInformation";
+import { CreditCardInformationParent } from "../CreditCardInformation";
+import { MessageParent } from "../Message";
+import { PricingParent } from "../Pricing";
+import { PlaceViewsParent } from "../PlaceViews";
+import { GuestRequirementsParent } from "../GuestRequirements";
+import { PoliciesParent } from "../Policies";
+import { HouseRulesParent } from "../HouseRules";
+import { AmenitiesParent } from "../Amenities";
+
+import { Context } from "./context";
 
 export interface Types extends ITypes {
   Context: Context;
